fix: adding if condition
