+++ src/pages/derivx/index.tsx
@@ -16,64 +16,64 @@ import { isBrowser } from 'common/utility'
 import useRegion from 'components/hooks/use-region'
 
 const DerivX = () => {
-    // const [is_mobile, setMobile] = useState(false)
-    // const handleResizeWindow = useCallback(() => {
-    //     setMobile(isBrowser() ? window.screen.width <= size.tablet : false)
-    // }, [setMobile])
+    const [is_mobile, setMobile] = useState(false)
+    const handleResizeWindow = useCallback(() => {
+        setMobile(isBrowser() ? window.screen.width <= size.tablet : false)
+    }, [setMobile])
 
     const { is_row } = useRegion()
-    // const [is_loaded, setLoaded] = useState(false)
+    const [is_loaded, setLoaded] = useState(false)
 
-    // useEffect(() => {
-    //     setLoaded(true)
-    // }, [])
+    useEffect(() => {
+        setLoaded(true)
+    }, [])
 
-    // useEffect(() => {
-    //     handleResizeWindow()
-    //     window.addEventListener('resize', handleResizeWindow)
+    useEffect(() => {
+        handleResizeWindow()
+        window.addEventListener('resize', handleResizeWindow)
 
-    //     return () => {
-    //         window.removeEventListener('resize', handleResizeWindow)
-    //     }
-    // }, [handleResizeWindow])
+        return () => {
+            window.removeEventListener('resize', handleResizeWindow)
+        }
+    }, [handleResizeWindow])
 
-    // if (is_loaded) {
-    if (is_row) {
-        return (
-            <Layout>
-                <SEO
-                    title={localize(
-                        'Deriv X - a multi-asset CFD trading platform available on Deriv',
-                    )}
-                    description={localize(
-                        'Deriv X is a fully customisable, easy-to-use online trading platform offering CFDs on forex, commodities, cryptocurrencies, and synthetic indices.',
-                    )}
-                />
-                <Hero />
-                <SellingPoints />
-                <WhatIsDeriv />
-                <WhyTradeDerivX />
-                <StartDerivX />
-                <Accounts />
-                <DerivXGetApp />
-                <OurPlatforms />
-            </Layout>
-        )
+    if (is_loaded) {
+        if (is_row) {
+            return (
+                <Layout>
+                    <SEO
+                        title={localize(
+                            'Deriv X - a multi-asset CFD trading platform available on Deriv',
+                        )}
+                        description={localize(
+                            'Deriv X is a fully customisable, easy-to-use online trading platform offering CFDs on forex, commodities, cryptocurrencies, and synthetic indices.',
+                        )}
+                    />
+                    <Hero />
+                    <SellingPoints />
+                    <WhatIsDeriv />
+                    <WhyTradeDerivX />
+                    <StartDerivX />
+                    <Accounts />
+                    <DerivXGetApp />
+                    <OurPlatforms />
+                </Layout>
+            )
+        }
+
+        return <PageNotFound />
     }
 
-    return <PageNotFound />
-    // }
-
-    // return (
-    //     <>
-    //         <SEO
-    //             title={localize('Deriv X - a multi-asset CFD trading platform available on Deriv')}
-    //             description={localize(
-    //                 'Deriv X is a fully customisable, easy-to-use online trading platform offering CFDs on forex, commodities, cryptocurrencies, and synthetic indices.',
-    //             )}
-    //         />
-    //     </>
-    // )
+    return (
+        <>
+            <SEO
+                title={localize('Deriv X - a multi-asset CFD trading platform available on Deriv')}
+                description={localize(
+                    'Deriv X is a fully customisable, easy-to-use online trading platform offering CFDs on forex, commodities, cryptocurrencies, and synthetic indices.',
+                )}
+            />
+        </>
+    )
 }
 
 export default WithIntl()(DerivX)
